backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-community
httpx
pydantic
langchain-openai
qiniu
sqlalchemy
pymysql
